requirements.txt
# Ross Cameron Screener Dependencies
yfinance>=0.2.18
pandas>=1.5.0
numpy>=1.24.0
requests>=2.28.0
alpaca-trade-api>=3.0.0
pandas-ta>=0.3.14b0
python-dotenv>=1.0.0
schedule>=1.2.0

# News and Sentiment Analysis
textblob>=0.17.1
alpha-vantage>=3.0.0
beautifulsoup4>=4.12.0
nltk>=3.8
vaderSentiment>=3.3.2
